Fix --without-components with subsetted components
It's not entirely clear what changed in 1.66, but rust-lang/rust#105755
shows that we are failing to run the install script with --without if
there are subsetted component names.

This changes the behavior of the filtering to require an *exact* match
rather than a partial match, which seems like the better way to go. It's
not very clear to me that the previous behavior was actually a good
idea.