Write out manifest.in entries deterministically
Currently, the order of the entries in `manifest.in` depends on the
iteration order of `copy_with_callback`, which in turn depends on
`WalkDir`, which explicitly says iteration order is unspecified. It's
possible to call `WalkDir::sort_by` to give an iteration order for each
directory's entries, but it felt better to accumulate the lines and then
sort them to a) make it more evident that this is happening, and b)
enable the copying to be parallelized in the future.